style single date picker to match mock
