+++ packages/augur-ui/src/modules/common/selection.tsx
@@ -3,10 +3,6 @@ import classNames from "classnames";
 import Styles from "modules/common/selection.styles";
 import { Chevron, DotDotDot, TwoArrows } from "modules/common/icons";
 
-import "react-dates/initialize";
-import "react-dates/lib/css/_datepicker.css";
-import { SingleDatePicker } from "react-dates";
-
 export interface NameValuePair {
   label: string;
   value: string | number;
@@ -46,37 +42,6 @@ interface PillSelectionProps {
 interface PillSelectionState {
   selected: number;
 }
-interface DatePickerProps {
-  id?: string;
-  date: any;
-  placeholder?: string;
-  onDateChange: Function;
-  isOutsideRange?: Function;
-  focused?: boolean;
-  onFocusChange?: Function;
-  displayFormat: string;
-  numberOfMonths: number;
-  navPrev?: any;
-  navNext?: any;
-}
-
-export const DatePicker = (props: DatePickerProps) => (
-  <div className={Styles.DatePicker}>
-    <SingleDatePicker
-      id={props.id}
-      date={props.date}
-      placeholder={props.placeholder || "Date (D MMM YYYY)"}
-      onDateChange={props.onDateChange}
-      isOutsideRange={props.isOutsideRange || (() => false)}
-      focused={props.focused}
-      onFocusChange={props.onFocusChange}
-      displayFormat={props.displayFormat || "D MMM YYYY"}
-      numberOfMonths={props.numberOfMonths}
-      navPrev={props.navPrev || Chevron}
-      navNext={props.navNext || Chevron}
-    />
-  </div>
-);
 
 interface DotSelectionProps {
   children: React.StatelessComponent;
