+++ packages/augur-ui/src/modules/market/components/market-header/market-header-reporting.styles.less
@@ -78,7 +78,7 @@
   }
 }
 
-.Break {
+.EmptyBreak {
   background: @color-theme-primary;
   height: 0.125rem;
